[ADD] unifrut Consignment Setting and Fields
Enabling the setting Consignment (`group_stock_tracking_owner`).
Adding in the Operation Type model the field `assign_owner_required` to
force setting an Owner in the Transfer where that operation type is
selected, in those cases the `partner_id` will be set as the 'Assign
Owner'. This field will also allow showing the field `price_unit` in the
`stock.move` tree view inside the `stock.picking` form view.

Also, adding 3 other fields for the settings 2 of them are Accounts and
the other will be the Journal, these will be used for accounting
consignment inventory arrivals and departures.

And adding a Cron that will create the Consignment Transfers once a day.

[FIX] unifrut: Missing Post-Install in Tests

As the tests in the promissory note file depend on the Mexican
Localization the tag `post_install` should be set because it requires a
CoA to be installed.
With this change, the following error in the pipeline will be avoided:
'This test requires a CoA to be installed, it should be tagged "post_install"'