[ADD] unifrut: "Stock Picking Return Restricted Qty" Module
Adding the module `stock_picking_return_restricted_qty` to the manifest,
this module will allow a validation that the quantity returned is not more
than the original quantity.

Also, a test to prove the behavior in the 3 possible picking types
"incoming", "outcoming" and "internal".

The tags `"post_install", "-at_install"` were added to the class of the
test to ensure the installation of the module before running the tests.