[FIX] unifrut: Get of "Early Payment Discount"
Before these changes, the only thing that was checked for, to apply or
not an early payment discount, was the number of months between the
invoice date and the payment date, and the first match was the one
applied.

Having two early payment discounts for the same month using pricelists as
the price calculation and in the first "match" of numbers of months
the product or date didn't apply was causing setting the `lst_price`
(which was $1 MXN in a USD pricelist so it was taken as a dollar and it
was converted to MXN.) instead of using the other pricelist that had the
correct price.

So, now if there is more than one early payment discount with a pricelist
it will check first if there is at least one item in that pricelist that
applies to the product.

[IMP] unifrut: Product Early Payment Discount Demo and Tests

Adding some demo data to allow doing some functional testing about early
payment discounts without having to manually create it each time a Demo
DB is created.

Also, adding some tests that were missing about "Early Payment
Discounts", one for each type of price computation (Percentage, Fixed
Amount, and Pricelist), one to check that when the default early payment
product is not configured as a parameter an error is raised, and one to
check that the creation of early payment discounts (Credit Notes) when
the invoice is not candidate and when there is no early payment discount
that applies to the product.