[ADD] unifrut: new user templates for managing user profiles
This creates user templates for use as user roles and also creates
new groups and user permissions. This in done to avoid giving all
users the admin role to access reports and other functionality.
This adds the functionality to sort groups by application as a list
This adds the following modules as dependencies as they are
required for the role permissions:
- l10n_mx_edi_cancellation
- hr_contract

The user templates created are:
- VENTAS CAJERO
- CRM VENDEDOR
- GERENTE SUCURSALES
- AUX. ALMACEN CUAUHTEMOC
- AUX. ALMACEN SUCURSAL
- ALMACEN
- CREDITO Y COBRANZA
- COBRANZA
- DIRECCION GENERAL
- GESTION, CONTROL Y SEGUIMIENTO
- INVENTARIOS
- COMPRAS
- RH
- CONTABILIDAD